refactor: Quax internal org
- Added `_dispatch.py` to handle multiple dispatch registration and resolution for JAX primitives.
- Introduced `_primitives.py` to define custom JAX primitives with Quax support, including `jit_quax`, `while_quax`, `cond_quax`, and `scan_quax`.
- Created `_quaxify.py` to wrap functions for Quax compatibility, enabling dynamic dispatch based on input types.
- Developed `_trace.py` to manage tracing of Quax operations, including custom JVP rules.
- Defined `_values.py` to establish the `Value` class hierarchy, facilitating multiple dispatch with custom array-like objects.
- Updated unit tests to reflect changes in cache management and ensure proper functionality of the new dispatch system.
