--- src/quax/_core.py
@@ -1,875 +1,0 @@
-import abc
-import functools as ft
-import itertools as it
-import weakref
-from collections.abc import Callable, Sequence
-from typing import (
-    Any,
-    cast,
-    Generic,
-    no_type_check,
-    overload,
-    TypeAlias,
-    TypeGuard,
-    TypeVar,
-    Union,
-)
-
-import equinox as eqx
-import jax
-import jax._src.core as core
-import jax.extend.core as jexc
-import jax.extend.linear_util as lu
-import jax.numpy as jnp
-import jax.tree_util as jtu
-import plum
-from jax.custom_derivatives import SymbolicZero as SZ
-from jaxtyping import ArrayLike, PyTree
-
-from ._compat import JAX_GE_0_9_2, jit_p, typeof
-
-
-T = TypeVar("T")
-CT = TypeVar("CT", bound=Callable)
-ValueLike: TypeAlias = Union[ArrayLike, "Value"]
-
-#
-# Rules
-#
-
-
-_rules: dict[jexc.Primitive, plum.Function] = {}
-
-# Cache resolved dispatch methods keyed by (primitive, arg-types).
-# Sentinel for "no matching rule; fall through to _default_process".
-_DISPATCH_MISS = object()
-_dispatch_cache: dict[tuple, Any] = {}
-
-
-def register(primitive: jexc.Primitive, *, precedence: int = 0) -> Callable[[CT], CT]:
-    """Registers a multiple dispatch implementation for this JAX primitive.
-
-    !!! Example
-
-        Used as decorator, and requires type annotations to perform multiple dispatch:
-        ```python
-        @quax.register(jax.lax.add_p)
-        def _(x: SomeValue, y: SomeValue):
-            return ...  # some implementation
-        ```
-
-    All positional arguments will be (subclasses of) [`quax.Value`][] -- these are the
-    set of types that Quax will attempt to perform multiple dispatch with.
-
-    All keyword arguments will be the parameters for this primitive, as passed to
-    `prim.bind(... **params)`.
-
-    **Arguments:**
-
-    - `primitive`: The `jax.extend.core.Primitive` to provide a multiple
-      dispatch implementation for.
-
-    - `precedence`: The precedence of this rule.
-        See `plum.Dispatcher.dispatch` for details.
-
-    **Returns:**
-
-    A decorator for registering a multiple dispatch rule with the specified primitive.
-    """
-
-    def _register(rule: CT) -> CT:
-        try:
-            existing_rule = _rules[primitive]
-        except KeyError:
-
-            def new_rule():
-                raise NotImplementedError("Abstract primitive")  # pragma: no cover
-
-            new_rule.__name__ = f"{primitive}_dispatcher"
-            new_rule.__qualname__ = f"{primitive}_dispatcher"
-            existing_rule = plum.Dispatcher().abstract(new_rule)
-
-            _rules[primitive] = existing_rule
-        existing_rule.dispatch(rule, precedence=precedence)
-        # Invalidate any cached dispatch decisions for this primitive so that
-        # newly registered rules are picked up on the next call.
-        keys_to_drop = [k for k in _dispatch_cache if k[0] is primitive]
-        for k in keys_to_drop:
-            del _dispatch_cache[k]
-        return rule
-
-    return _register
-
-
-#
-# Interpreter
-#
-
-
-class _QuaxTracer(core.Tracer):
-    __slots__ = ("value",)
-
-    def __init__(self, trace: "_QuaxTrace", value: "Value") -> None:
-        assert _is_value(value)
-        self._trace = trace
-        self.value = value
-
-    @property
-    def aval(self) -> core.AbstractValue:  # pyright: ignore[reportIncompatibleVariableOverride]
-        v = self.value
-        # Fast path for _DenseArrayValue: bypass eqx's __getattribute__ which
-        # wraps every method access in a new BoundMethod (another eqx.Module).
-        # JAX calls .aval on every tracer throughout tracing, so this is hot.
-        if type(v) is _DenseArrayValue:
-            return typeof(object.__getattribute__(v, "array"))
-        return v.aval()
-
-    def full_lower(self) -> Union[ArrayLike, "_QuaxTracer"]:
-        return (
-            core.full_lower(self.value.array)  # pyright: ignore[reportAttributeAccessIssue]
-            if isinstance(self.value, _DenseArrayValue)
-            else self
-        )
-
-
-DefaultCallable: TypeAlias = Callable[
-    [jexc.Primitive, Sequence[ValueLike], dict[str, Any]],
-    ValueLike | Sequence[ValueLike],
-]
-
-
-def _default_process(
-    primitive: jexc.Primitive, values: Sequence[ValueLike], params: dict[str, Any]
-) -> ValueLike | Sequence[ValueLike]:
-    # Fast path: find first non-default, then check if all match
-    default: DefaultCallable | None = None
-    value_default = Value.default  # Cache attribute lookup
-    for x in values:
-        if isinstance(x, Value):
-            x_default = type(x).default
-            if x_default is value_default:
-                continue
-            if default is None:
-                default = x_default
-            elif default is not x_default:
-                # Multiple different defaults - slow path
-                types = {type(v) for v in values if isinstance(v, Value)}
-                raise TypeError(
-                    f"Multiple array-ish types {types} are specifying default "
-                    f"process rules."
-                )
-        elif not eqx.is_array_like(x):
-            assert False
-
-    if default is None:
-        default = value_default
-
-    return default(primitive, values, params)
-
-
-def _wrap_if_array(x: Union[ArrayLike, "Value"]) -> "Value":
-    return (
-        _DenseArrayValue(cast(ArrayLike, x)) if eqx.is_array_like(x) else cast(Value, x)
-    )
-
-
-class _QuaxTrace(
-    core.Trace if JAX_GE_0_9_2 else core.Trace[_QuaxTracer],  # pyright: ignore[reportGeneralTypeIssues, reportInvalidTypeArguments]
-):
-    __slots__ = ("tag", "parent_trace")
-
-    def __init__(self, parent_trace: core.Trace | None, tag: core.TraceTag) -> None:
-        self.tag = tag
-        self.parent_trace = parent_trace
-        super().__init__()
-
-    def to_value(self, val: Any) -> Any:
-        if isinstance(val, _QuaxTracer) and val._trace.tag is self.tag:  # type: ignore[attr-defined]
-            return val.value
-        return _DenseArrayValue(val)
-
-    # ===========================================
-    # Override methods from jax.core.Trace
-
-    def process_primitive(
-        self, primitive: jexc.Primitive, tracers: Sequence[Any], params: dict[str, Any]
-    ) -> _QuaxTracer | list[_QuaxTracer]:
-        """Processes a primitive with the given tracers and parameters.
-
-        This is the main entry point for processing primitives in JAX. It is
-        called when a primitive is encountered during the evaluation of a JAX
-        expression.
-
-        **Arguments:**
-        - `primitive`: The primitive to be processed.
-        - `tracers`: A sequence of tracers that represent the inputs to the
-          primitive.
-        - `params`: A dictionary of parameters for the primitive.
-
-        **Returns:**
-        A ``_QuaxTracer`` (or list thereof) representing the result of
-        processing the primitive.
-
-        """
-        # ── O1: fast path for all-dense inputs with no registered rule ──────
-        # When every input is one of *our* _QuaxTracer wrappers around a plain
-        # JAX array (_DenseArrayValue) and there is no registered quax rule for
-        # this primitive, skip the entire dispatch stack and delegate directly
-        # to the parent trace.  This avoids per-call allocations of
-        # _DenseArrayValue objects, the plum lookup, _default_process, and the
-        # materialise loop — the common case when quaxify wraps pure-JAX code.
-        if primitive not in _rules:
-            tag = self.tag
-            arrays: list[Any] = []
-            for t in tracers:
-                if not (isinstance(t, _QuaxTracer) and t._trace.tag is tag):  # type: ignore[attr-defined]
-                    break
-                v = t.value
-                if not isinstance(v, _DenseArrayValue):
-                    break
-                arrays.append(v.array)
-            else:
-                # All inputs were our dense tracers
-                with core.set_current_trace(self.parent_trace):
-                    out = primitive.bind(*arrays, **params)
-                if primitive.multiple_results:
-                    return [_QuaxTracer(self, _DenseArrayValue(x)) for x in out]  # pyright: ignore[reportGeneralTypeIssues]
-                return _QuaxTracer(self, _DenseArrayValue(out))  # pyright: ignore[reportArgumentType]
-
-        # ── full dispatch path ───────────────────────────────────────────────
-        # Parse the tracers into values, unpacking any _DenseArrayValues.
-        values = tuple(
-            (x.array if isinstance(x := self.to_value(t), _DenseArrayValue) else x)
-            for t in tracers
-        )
-
-        # ── O2: cache resolved dispatch method per (primitive, arg-types) ───
-        # plum's resolve_method performs MRO inspection on every call; cache
-        # the result so repeated invocations with the same type signature pay
-        # the lookup cost only once.
-        cache_key = (primitive, tuple(type(v) for v in values))
-        cached = _dispatch_cache.get(cache_key)
-
-        with core.set_current_trace(self.parent_trace):
-            if cached is _DISPATCH_MISS:
-                out = _default_process(primitive, values, params)
-            elif cached is not None:
-                out = cached(*values, **params)
-            elif (rule := _rules.get(primitive)) is None:
-                # First time seeing this (primitive, types) combination.
-                _dispatch_cache[cache_key] = _DISPATCH_MISS
-                out = _default_process(primitive, values, params)
-            else:
-                try:
-                    method, _ = rule.resolve_method(values)
-                except plum.NotFoundLookupError:
-                    _dispatch_cache[cache_key] = _DISPATCH_MISS
-                    out = _default_process(primitive, values, params)
-                else:
-                    _dispatch_cache[cache_key] = method
-                    out = method(*values, **params)
-
-        # Post-process the output
-        if primitive.multiple_results:
-            return [_QuaxTracer(self, _wrap_if_array(x)) for x in out]  # pyright: ignore[reportGeneralTypeIssues]
-        return _QuaxTracer(self, _wrap_if_array(out))  # pyright: ignore[reportArgumentType]
-
-    if JAX_GE_0_9_2:
-        # In JAX v0.9.2+ (PR https://github.com/jax-ml/jax/pull/35730) JAX was
-        # changed to have an extra argument -- `avals` -- in bind_with_trace
-        # before params.
-
-        def process_custom_jvp_call(
-            self, primitive, fun, jvp, tracers, *, symbolic_zeros
-        ) -> list[_QuaxTracer]:
-            # Each `t.value` will be some `Value`, and thus a PyTree. Here we
-            # flatten the `Value`-ness away.
-            in_leaves, in_treedef = jtu.tree_flatten(
-                [self.to_value(t) for t in tracers]
-            )
-            fun, out_treedef1 = _custom_jvp_fun_wrap(fun, self.tag, in_treedef)
-            jvp, out_treedef2 = _custom_jvp_jvp_wrap(jvp, self.tag, in_treedef)
-            avals = tuple(x.aval if type(x) is SZ else typeof(x) for x in in_leaves)
-            params = dict(subfuns=(fun, jvp), symbolic_zeros=symbolic_zeros)
-            out_leaves = primitive.bind_with_trace(
-                self.parent_trace, tuple(in_leaves), avals, params
-            )
-            _, out_treedef = lu.merge_linear_aux(out_treedef1, out_treedef2)
-            out_values = jtu.tree_unflatten(out_treedef, out_leaves)
-            return [_QuaxTracer(self, x) for x in out_values]
-
-    else:
-
-        def process_custom_jvp_call(
-            self, primitive, fun, jvp, tracers, *, symbolic_zeros
-        ) -> list[_QuaxTracer]:
-            # Each `t.value` will be some `Value`, and thus a PyTree. Here we
-            # flatten the `Value`-ness away.
-            in_leaves, in_treedef = jtu.tree_flatten(
-                [self.to_value(t) for t in tracers]
-            )
-            fun, out_treedef1 = _custom_jvp_fun_wrap(fun, self.tag, in_treedef)
-            jvp, out_treedef2 = _custom_jvp_jvp_wrap(jvp, self.tag, in_treedef)
-            out_leaves = primitive.bind_with_trace(
-                self.parent_trace,
-                (fun, jvp, *in_leaves),
-                dict(symbolic_zeros=symbolic_zeros),
-            )
-            _, out_treedef = lu.merge_linear_aux(out_treedef1, out_treedef2)
-            out_values = jtu.tree_unflatten(out_treedef, out_leaves)
-            return [_QuaxTracer(self, x) for x in out_values]
-
-    # TODO: add other process_* rules
-
-
-@lu.transformation_with_aux
-def _custom_jvp_fun_wrap(tag, in_treedef, *in_leaves):
-    in_values = jtu.tree_unflatten(in_treedef, in_leaves)
-    with core.take_current_trace() as parent_trace:
-        trace = _QuaxTrace(parent_trace, tag)
-        in_tracers = [x if type(x) is SZ else _QuaxTracer(trace, x) for x in in_values]
-        with core.set_current_trace(trace):
-            out_tracers = yield in_tracers, {}
-            # The symbolic zero branch here will actually create a `quax.zero.Zero`!
-            out_tracers = [
-                jnp.zeros(t.aval.shape, t.aval.dtype) if type(t) is SZ else t  # pyright: ignore[reportAttributeAccessIssue]
-                for t in out_tracers
-            ]
-            out_values = [trace.to_value(t) for t in out_tracers]
-            del out_tracers
-        del trace, in_tracers
-    out_leaves, out_treedef = jtu.tree_flatten(out_values)
-    yield out_leaves, out_treedef
-
-
-@lu.transformation_with_aux
-def _custom_jvp_jvp_wrap(tag, in_treedef, *in_primals_and_tangents):
-    split = len(in_primals_and_tangents) // 2
-    in_primal_values = jtu.tree_unflatten(in_treedef, in_primals_and_tangents[:split])
-    in_tangent_values_raw = jtu.tree_unflatten(
-        in_treedef, in_primals_and_tangents[split:]
-    )
-    # When symbolic_zeros=True, JAX may pass SymbolicZero tangent leaves. After
-    # unflattening, SZs can be embedded inside a Value (e.g. MyArray(SZ)),
-    # breaking .aval() calls. Promote only fully-symbolic tangents back to a
-    # value-level SZ so the JVP rule can check `type(t) is SZ` directly, while
-    # leaving mixed tangents untouched.
-    in_tangent_values = [
-        SZ(p.aval())
-        if (leaves := jtu.tree_leaves(t)) and all(type(l) is SZ for l in leaves)
-        else t
-        for p, t in zip(in_primal_values, in_tangent_values_raw)
-    ]
-    # Calling `_QuaxTracer` directly here, not using `trace.{pure,lift}` as each `x` is
-    # a `Value`, not an array (=> pure) or tracer (=> lift).
-    with core.take_current_trace() as parent_trace:
-        trace = _QuaxTrace(parent_trace, tag)
-        in_tracers = [
-            x if type(x) is SZ else _QuaxTracer(trace, x)
-            for x in it.chain(in_primal_values, in_tangent_values)
-        ]
-        with core.set_current_trace(trace):
-            out_tracers = yield in_tracers, {}
-            # The symbolic zero branch here will actually create a `quax.zero.Zero`!
-            out_tracers = [
-                jnp.zeros(t.aval.shape, t.aval.dtype) if type(t) is SZ else t  # pyright: ignore[reportAttributeAccessIssue]
-                for t in out_tracers
-            ]
-            out_values = [trace.to_value(t) for t in out_tracers]
-            # Pre-calculate split point
-            out_split = len(out_values) // 2
-            out_primal_values = out_values[:out_split]
-            out_tangent_values = out_values[out_split:]
-            out_primal_values2 = []
-            out_tangent_values2 = []
-            for primal, tangent in zip(
-                out_primal_values, out_tangent_values, strict=True
-            ):
-                if primal.__class__ != tangent.__class__:
-                    primal = primal.materialise()
-                    tangent = tangent.materialise()
-                out_primal_values2.append(primal)
-                out_tangent_values2.append(tangent)
-            del out_tracers
-        del trace, in_tracers
-    out_primals, out_primal_treedef = jtu.tree_flatten(out_primal_values2)
-    out_tangents, out_tangent_treedef = jtu.tree_flatten(out_tangent_values2)
-    if out_primal_treedef != out_tangent_treedef:
-        raise ValueError(
-            "Primals and tangents had the same class, but different flattened results."
-        )
-    yield out_primals + out_tangents, out_primal_treedef
-
-
-#
-# API
-#
-
-
-# Any -> Any so overloads carry the public types. mypy can't prove the else
-# branch is T (since T may be Value). To type the body, use Union[Value, T]
-# + cast(T, x), or constrain T to exclude Value.
-@overload
-def _wrap_tracer(trace: _QuaxTrace, x: "Value") -> _QuaxTracer: ...
-@overload
-def _wrap_tracer(trace: _QuaxTrace, x: T) -> T: ...
-def _wrap_tracer(trace: _QuaxTrace, x: Any) -> Any:
-    return _QuaxTracer(trace, x) if _is_value(x) else x
-
-
-def _unwrap_tracer(trace: "_QuaxTrace", x: Any) -> Any:
-    if eqx.is_array_like(x):
-        x = trace.full_raise(x)
-    if isinstance(x, _QuaxTracer):
-        return x.value.array if isinstance(x.value, _DenseArrayValue) else x.value
-    return x
-
-
-class _Quaxify(eqx.Module, Generic[CT]):
-    fn: CT
-    filter_spec: PyTree[bool | Callable[[Any], bool]]
-    dynamic: bool = eqx.field(static=True)
-
-    @property
-    def __wrapped__(self) -> CT:
-        return self.fn
-
-    def __call__(self, *args: Any, **kwargs: Any) -> Any:
-        dynamic, static = eqx.partition(
-            (self.fn, args, kwargs), self.filter_spec, is_leaf=_is_value
-        )
-        tag = core.TraceTag()
-        with core.take_current_trace() as parent_trace:
-            trace = _QuaxTrace(parent_trace, tag)
-            # Cache partial functions to avoid repeated closure creation
-            dynamic = jtu.tree_map(
-                ft.partial(_wrap_tracer, trace), dynamic, is_leaf=_is_value
-            )
-            fn, args, kwargs = eqx.combine(dynamic, static)
-            with core.set_current_trace(trace):
-                out = fn(*args, **kwargs)
-            out = jtu.tree_map(ft.partial(_unwrap_tracer, trace), out)
-            return out
-
-    def __get__(
-        self, instance: object | None, owner: Any
-    ) -> Union["_Quaxify[CT]", eqx.Partial["_Quaxify[CT]"]]:
-        # Getting from a class
-        if instance is None:
-            return self
-        # Getting from an instance
-        return eqx.Partial(self, instance)
-
-
-def quaxify(
-    fn: CT,
-    filter_spec: PyTree[bool | Callable[[Any], bool]] = True,
-) -> CT:
-    """'Quaxifies' a function, so that it understands custom array-ish objects like
-    [`quax.examples.lora.LoraArray`][]. When this function is called, multiple dispatch
-    will be performed against the types it is called with.
-
-    **Arguments:**
-
-    - `fn`: the function to wrap.
-    - `filter_spec`: which arguments to quaxify. Advanced usage, see tip below.
-
-    **Returns:**
-
-    A copy of `fn`, that understands all Quax types.
-
-    !!! Tip "Only quaxifying some argments"
-
-        Calling `quax.quaxify(fn, filter_spec)(*args, **kwargs)` will under-the-hood run
-        `dynamic, static = eqx.partition((fn, args, kwargs), filter_spec)`, and then
-        only quaxify those arguments in `dynamic`. This allows for passing through some
-        [`quax.Value`][]s into the function unchanged, typically so that they can hit a
-        nested `quax.quaxify`. See the
-        [advanced tutorial](../examples/redispatch.ipynb).
-    """
-    return cast(CT, eqx.module_update_wrapper(_Quaxify(fn, filter_spec, dynamic=False)))
-
-
-#
-# Values
-#
-
-
-class Value(eqx.Module):
-    """Represents an object which Quax can perform multiple dispatch with.
-
-    In practice you will almost always want to inherit from [`quax.ArrayValue`][]
-    instead, which represents specifically an array-ish object that can be used for
-    multiple dispatch.
-    """
-
-    @abc.abstractmethod
-    def aval(self) -> core.AbstractValue:
-        """All concrete subclasses must implement this method, specifying the abstract
-        value seen by JAX.
-
-        **Arguments:**
-
-        Nothing.
-
-        **Returns:**
-
-        Any subclass of `jax.core.AbstractValue`. Typically a `jax.core.ShapedArray`.
-        """
-
-    @staticmethod
-    def default(
-        primitive: jexc.Primitive, values: Sequence[ValueLike], params: dict[str, Any]
-    ) -> ValueLike | Sequence[ValueLike]:
-        """This is the default rule for when no rule has been [`quax.register`][]'d for
-        a primitive.
-
-        When performing multiple dispatch `primitive.bind(value1, value2, value3)`,
-        then:
-
-        1. If there is a dispatch rule matching the types of `value1`, `value2`, and
-            `value3`, then that will be used.
-        2. If precisely one of the types of `value{1,2,3}` overloads this method, then
-            that default rule will be used.
-        3. If precisely zero of the types of `value{1,2,3}` overloads this method, then
-            all values are [`quax.Value.materialise`][]d, and the usual JAX
-            implementation is called.
-        4. If multiple of the types of `value{1,2,3}` overload this method, then a
-            trace-time error will be raised.
-
-        **Arguments:**
-
-        - `primitive`: the `jax.extend.core.Primitive` being considered.
-        - `values`: a sequence of what values this primitive is being called with. Each
-            value can either be [`quax.Value`][]s, or a normal JAX arraylike (i.e.
-            `bool`/`int`/`float`/`complex`/NumPy scalar/NumPy array/JAX array).
-        - `params`: the keyword parameters to the primitive.
-
-        **Returns:**
-
-        The result of binding this primitive against these types. If
-        `primitive.multiple_results is False` then this should be a single `quax.Value`
-        or JAX arraylike. If `primitive.multiple_results is True`, then this should be
-        a tuple/list of such values.
-
-        !!! Example
-
-            The default implementation discussed above performs the following:
-            ```python
-            @staticmethod
-            def default(primitive, values, params):
-                arrays = [
-                    x if equinox.is_array_like(x) else x.materialise() for x in values
-                ]
-                return primitive.bind(*arrays, **params)
-            ```
-            (Using the [Equinox](https://github.com/patrick-kidger/equinox) library that
-            underlies much of the JAX ecosystem.)
-        """
-        # Use list comprehension for better performance
-        arrays = [
-            x.materialise() if _is_value(x) else cast(ArrayLike, x) for x in values
-        ]
-        return primitive.bind(*arrays, **params)
-
-    @abc.abstractmethod
-    def materialise(self) -> Any:
-        """All concrete subclasses must implement this method, specifying how to
-        materialise this object into a JAX type (i.e. almost always a JAX array, unless
-        you're doing something obscure using tokens or refs).
-
-        !!! Example
-
-            For example, a LoRA array consists of three arrays `(W, A, B)`, combined as
-            `W + AB`. [`quax.examples.lora.LoraArray`] leaves these as three separate
-            arrays for efficiency, but calling `lora_array.materialise()` will evaluate
-            `W + AB` and return a normal JAX array.
-
-        This is so that the usual JAX primitive implementations can be applied as a
-        fallback: the array-ish object is materialised, and then the usual JAX
-        implementation called on it. (See [`quax.Value.default`][].)
-
-        !!! Info
-
-            It is acceptable for this function to just raise an error -- in this case
-            the error will be surfaced to the end user, indicating that an operation is
-            not supported for this array-ish object.
-
-        **Arguments:**
-
-        Nothing.
-
-        **Returns:**
-
-        A JAX type; typically a JAX array.
-        """
-
-
-def _is_value(x: object) -> TypeGuard[Value]:
-    return isinstance(x, Value)
-
-
-class ArrayValue(Value):
-    """A subclass [`quax.Value`][] for specifically array-like types. If you are
-    creating a custom array-ish object then you should typically inherit from this.
-
-    Provides the properties `.shape`, `.dtype`, `.ndim`, `.size`, each as a shortcut for
-    `self.aval().shape` etc.
-    """
-
-    @abc.abstractmethod
-    def materialise(self) -> ArrayLike:
-        pass
-
-    @abc.abstractmethod
-    def aval(self) -> core.ShapedArray:
-        pass
-
-    @property
-    def shape(self) -> tuple[int, ...]:
-        return self.aval().shape
-
-    @property
-    def dtype(self) -> Any:  # jax.numpy.dtype or ExtendedDType
-        return self.aval().dtype
-
-    @property
-    def ndim(self) -> int:
-        return self.aval().ndim
-
-    @property
-    def size(self) -> int:
-        return self.aval().size
-
-
-class _DenseArrayValue(ArrayValue):
-    """Internal type used to wrap up a JAX arraylike into Quax's `Value` system.
-
-    This is an implementation detail hidded from the user! It is unwrapped straight
-    before calling a dispatch rule, and re-wrapped immediately afterwards.
-    """
-
-    array: ArrayLike
-
-    def materialise(self) -> ArrayLike:
-        return self.array
-
-    def aval(self) -> core.ShapedArray:
-        return typeof(self.array)
-
-
-def _make_cache_finalizer(cache: dict, key: tuple) -> Callable[[Any], None]:
-    """Return a weakref finalizer that removes *key* from *cache* when called."""
-
-    def _fin(_ref: Any) -> None:
-        cache.pop(key, None)
-
-    return _fin
-
-
-# Cache for jit_quax: maps (id(jaxpr), inline, treedef) -> (jaxpr_wref,
-# jitted_fn). entry[0] is a weakref to the jaxpr; the finalizer evicts the
-# entry when the jaxpr is GC'd. This prevents unbounded growth in programs that
-# generate many distinct jaxprs (e.g. dynamic code-gen).
-_jit_quax_cache: dict[tuple, tuple[Any, Any]] = {}
-
-
-@register(jit_p)
-def jit_quax(
-    *args: ArrayLike | ArrayValue, jaxpr: Any, inline: bool, **kwargs: Any
-) -> Any:
-    del kwargs
-
-    leaves, treedef = jtu.tree_flatten(args)  # remove all Values
-
-    # Without caching, every call constructs a fresh quaxify wrapper +
-    # lambda and calls jax.jit() on it.  jax.jit identifies functions by
-    # object identity, so a new lambda == a new compilation on every
-    # invocation — the dominant cost for functions like jnp.histogram2d
-    # that contain many internal @jit-decorated helpers.
-    #
-    # The jaxpr is stable across repeated calls with the same argument
-    # avals, so id(jaxpr) is a reliable key.  We also key by (inline,
-    # treedef) to handle the inline branch and structural arg differences.
-    key = (id(jaxpr), inline, treedef)
-    entry = _jit_quax_cache.get(key)
-    if entry is None:
-        fun = quaxify(jexc.jaxpr_as_fun(jaxpr))
-        wref = weakref.ref(jaxpr, _make_cache_finalizer(_jit_quax_cache, key))
-        if inline:  # For inline, store a plain callable; no jax.jit wrapper.
-            entry = (wref, fun)
-        else:
-            # Calling _Quaxify.__call__ directly (unbound) bypasses
-            # eqx.Module.__call__'s dir() + BoundMethod overhead.
-            _fun = cast(_Quaxify, fun)
-            qfun = lambda x: _Quaxify.__call__(_fun, *jtu.tree_unflatten(treedef, x))
-            jitted = jax.jit(qfun)
-            entry = (wref, jitted)
-        _jit_quax_cache[key] = entry
-
-    if inline:
-        return entry[1](*args)
-    return entry[1](leaves)  # now we can call without Quax.
-
-
-# Cache for while_quax: (id(cond_jaxpr), id(body_jaxpr), val_treedef)
-#   -> (cond_wref, body_wref, quax_cond_jaxpr, quax_body_jaxpr, val_treedef)
-# entry[0]/entry[1] are weakrefs; their finalizers evict the entry when either
-# original jaxpr is collected, preventing unbounded growth.
-_while_quax_cache: dict[tuple, tuple] = {}
-
-
-@register(jax.lax.while_p)
-def while_quax(
-    *args: ArrayValue | ArrayLike,
-    cond_nconsts: int,
-    cond_jaxpr: core.ClosedJaxpr,
-    body_nconsts: int,
-    body_jaxpr: core.ClosedJaxpr,
-) -> tuple[ArrayValue | ArrayLike, ...]:
-    body_end = cond_nconsts + body_nconsts
-    cond_consts = args[:cond_nconsts]
-    body_consts = args[cond_nconsts:body_end]
-    init_vals = args[body_end:]
-
-    cond_leaves, _ = jtu.tree_flatten(cond_consts)
-    body_leaves, _ = jtu.tree_flatten(body_consts)
-    init_val_leaves, val_treedef = jtu.tree_flatten(init_vals)
-
-    key = (id(cond_jaxpr), id(body_jaxpr), val_treedef)
-    entry = _while_quax_cache.get(key)
-    if entry is None:
-        quax_cond_fn = quaxify(jexc.jaxpr_as_fun(cond_jaxpr))
-        quax_cond_jaxpr = jax.make_jaxpr(quax_cond_fn)(*cond_consts, *init_vals)
-        quax_body_fn = quaxify(jexc.jaxpr_as_fun(body_jaxpr))
-        quax_body_jaxpr = jax.make_jaxpr(quax_body_fn)(*body_consts, *init_vals)
-        fin = _make_cache_finalizer(_while_quax_cache, key)
-        entry = (
-            weakref.ref(cond_jaxpr, fin),
-            weakref.ref(body_jaxpr, fin),
-            quax_cond_jaxpr,
-            quax_body_jaxpr,
-            val_treedef,
-        )
-        _while_quax_cache[key] = entry
-    else:
-        _, _, quax_cond_jaxpr, quax_body_jaxpr, val_treedef = entry
-
-    out_val = jax.lax.while_p.bind(
-        *cond_leaves,
-        *body_leaves,
-        *init_val_leaves,
-        cond_nconsts=cond_nconsts,
-        cond_jaxpr=quax_cond_jaxpr,
-        body_nconsts=body_nconsts,
-        body_jaxpr=quax_body_jaxpr,
-    )
-    result = jtu.tree_unflatten(val_treedef, out_val)
-    return result
-
-
-_sentinel = object()
-
-
-@register(jax.lax.cond_p)
-def cond_quax(
-    index: ArrayLike,
-    *args: ArrayValue | ArrayLike,
-    branches: tuple[core.ClosedJaxpr, ...],
-    linear: tuple[bool, ...] | object = _sentinel,
-    branches_platforms: tuple[str, ...] | object = _sentinel,
-) -> Any:
-    flat_args, in_tree = jtu.tree_flatten(args)
-
-    out_trees: list[Any] = []
-
-    def _make_quax_branch(jaxpr: core.ClosedJaxpr, /) -> core.ClosedJaxpr:
-        def flat_quax_call(flat_args: list[Any]) -> list[Any]:
-            _args = jtu.tree_unflatten(in_tree, flat_args)
-            flat_out, out_tree = jtu.tree_flatten(
-                quaxify(jexc.jaxpr_as_fun(jaxpr))(*_args)
-            )
-            out_trees.append(out_tree)
-            return flat_out
-
-        return jax.make_jaxpr(flat_quax_call)(flat_args)
-
-    quax_branches = tuple(_make_quax_branch(j) for j in branches)
-
-    if any(t != out_trees[0] for t in out_trees[1:]):
-        raise TypeError("all branches output must have the same pytree.")
-
-    kwargs = {"linear": linear} if linear is not _sentinel else {}
-    if branches_platforms is not _sentinel:
-        kwargs["branches_platforms"] = branches_platforms
-
-    out_val = jax.lax.cond_p.bind(index, *flat_args, branches=quax_branches, **kwargs)
-    return jtu.tree_unflatten(out_trees[0], out_val)
-
-
-# Cache for scan_quax: (id(jaxpr), consts_treedef, carry_treedef, xs_treedef)
-#   -> (jaxpr_wref, quax_jaxpr, out_treedef, nc, nv)
-# nc/nv = number of flat consts/carry leaves.
-# entry[0] is a weakref; the finalizer evicts the entry when the jaxpr is
-# collected, preventing unbounded growth.
-_scan_quax_cache: dict[tuple, tuple] = {}
-
-
-@register(jax.lax.scan_p)
-def scan_quax(
-    *args: ArrayValue | ArrayLike, num_consts: int, num_carry: int, jaxpr, **kwargs: Any
-) -> Any:
-    """Quax handler for ``lax.scan_p``.
-
-    Splits the flat ``args`` sequence into the three groups that ``lax.scan``
-    uses — constants, initial carry, and stacked ``xs`` — then builds a
-    quaxified jaxpr for the scan body and re-binds the primitive with it.
-
-    The quaxified body jaxpr is cached by ``(id(jaxpr), consts_treedef,
-    carry_treedef, xs_treedef)`` so that repeated calls with the same scan
-    body and pytree structure skip the ``jax.make_jaxpr`` tracing step.
-    """
-    consts_flat, c_tree = jtu.tree_flatten(args[:num_consts])
-    carry_flat, v_tree = jtu.tree_flatten(args[num_consts : num_consts + num_carry])
-    xs_flat, x_tree = jtu.tree_flatten(args[num_consts + num_carry :])
-
-    nc = len(consts_flat)
-    nv = len(carry_flat)
-
-    key = (id(jaxpr), c_tree, v_tree, x_tree)
-    entry = _scan_quax_cache.get(key)
-    if entry is None:
-        trace_in = (*consts_flat, *carry_flat, *[x[0, ...] for x in xs_flat])
-        fn = core.jaxpr_as_fun(jaxpr)
-
-        @no_type_check  # for beartype
-        def quax_fn(*flat: ArrayValue | ArrayLike) -> Any:
-            consts = jtu.tree_unflatten(c_tree, flat[:nc])
-            carry = jtu.tree_unflatten(v_tree, flat[nc : nc + nv])
-            xs = jtu.tree_unflatten(x_tree, flat[nc + nv :])
-            return quaxify(fn)(*consts, *carry, *xs)
-
-        quax_jaxpr, out_shape = jax.make_jaxpr(quax_fn, return_shape=True)(*trace_in)
-        out_tree = jtu.tree_structure(out_shape)
-        entry = (
-            weakref.ref(jaxpr, _make_cache_finalizer(_scan_quax_cache, key)),
-            quax_jaxpr,
-            out_tree,
-            nc,
-            nv,
-        )
-        _scan_quax_cache[key] = entry
-    else:
-        _, quax_jaxpr, out_tree, nc, nv = entry
-
-    out_flat = jax.lax.scan_p.bind(
-        *consts_flat,
-        *carry_flat,
-        *xs_flat,
-        jaxpr=quax_jaxpr,
-        num_consts=nc,
-        num_carry=nv,
-        **kwargs,
-    )
-
-    return jtu.tree_unflatten(out_tree, out_flat)
